+++ EnvScreen.c
@@ -33,7 +33,7 @@ void EnvScreen_delete(Object* this) {
 }
 
 void EnvScreen_draw(InfoScreen* this) {
-   InfoScreen_drawTitled(this, "Environment of process %d - %s", this->process->pid, this->process->comm);
+   InfoScreen_drawTitled(this, "Environment of process %d - %s", (int)this->process->pid, this->process->comm);
 }
 
 void EnvScreen_scan(InfoScreen* this) {
